- moved Qt.py in main directory
